requirements.txt
flask
flask-cors
pyrtlsdr
pyrtlsdrlib
numpy
scipy
matplotlib